e2e tests: move the localnet utils into the multi-homing pkg
